Task solution via Hibernate. v.1.0.1 Updated according comments: - delete user method realised via HQL - getAllUsers method realised via HQL - UserDao : universal object name for services - transaction rollback for DML-methods in case exceptions - other minor changes

+++ src/main/java/jm/task/core/jdbc/Main.java
@@ -24,7 +24,7 @@ public class Main {
         for (User u : usersList) {
             System.out.println(u);
         }
-        userService.removeUserById((long) 2);
+        userService.removeUserById(2L);
         userService.cleanUsersTable();
     }
 }
